pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "privastream"
dynamic = ["version"]
description = "AI-Powered Privacy Streaming Platform"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Privastream Team", email = "team@privastream.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.7.0",
    "numpy>=1.21.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "easyocr>=1.7.0",
    "insightface>=0.7.0",
    "onnxruntime>=1.15.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "Pillow>=9.5.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.910",
    "pre-commit>=2.15.0",
]
gpu = [
    "onnxruntime-gpu>=1.15.0",
]
audio = [
    "faster-whisper>=0.9.0",
    "librosa>=0.9.0", 
    "soundfile>=0.12.0",
    "webrtcvad>=2.0.10",
]

[project.scripts]
privastream = "privastream.cli.main:main"

[project.urls]
Homepage = "https://github.com/privastream/tiktok-techjam-2025"
Documentation = "https://privastream.readthedocs.io"
Repository = "https://github.com/privastream/tiktok-techjam-2025"
Issues = "https://github.com/privastream/tiktok-techjam-2025/issues"

[tool.setuptools_scm]
version_scheme = "post-release"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["privastream/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=privastream --cov-report=html --cov-report=term-missing"